Revert PR-4143: PartitionReassignmentHandler should only generate event when znode is created
RB=1541485
G=Kafka-Code-Reviews
R=okaraman
A=okaraman

+++ core/src/main/scala/kafka/controller/KafkaController.scala
@@ -901,29 +901,29 @@ class KafkaController(val config: KafkaConfig, zkClient: KafkaZkClient, time: Ti
    * method is invoked to avoid premature deletion of the `reassign_partitions` znode.
    */
   private def removePartitionsFromReassignedPartitions(partitionsToBeRemoved: Set[TopicPartition]) {
-    partitionsToBeRemoved.map(controllerContext.partitionsBeingReassigned).foreach { reassignContext =>
-      reassignContext.unregisterReassignIsrChangeHandler(zkClient)
-    }
-
-    val updatedPartitionsBeingReassigned = controllerContext.partitionsBeingReassigned -- partitionsToBeRemoved
-
-    info(s"Removing partitions $partitionsToBeRemoved from the list of reassigned partitions in zookeeper")
-
-    // write the new list to zookeeper
-    if (updatedPartitionsBeingReassigned.isEmpty) {
-      info(s"No more partitions need to be reassigned. Deleting zk path ${ReassignPartitionsZNode.path}")
-      zkClient.deletePartitionReassignment(controllerContext.epochZkVersion)
-      // Ensure we detect future reassignments
-      eventManager.put(PartitionReassignment)
-    } else {
-      val reassignment = updatedPartitionsBeingReassigned.mapValues(_.newReplicas)
-      try zkClient.setOrCreatePartitionReassignment(reassignment, controllerContext.epochZkVersion)
-      catch {
-        case e: KeeperException => throw new AdminOperationException(e)
+    if (partitionsToBeRemoved.nonEmpty) {
+      partitionsToBeRemoved.map(controllerContext.partitionsBeingReassigned).foreach { reassignContext =>
+        reassignContext.unregisterReassignIsrChangeHandler(zkClient)
       }
-    }
 
-    controllerContext.partitionsBeingReassigned --= partitionsToBeRemoved
+      val updatedPartitionsBeingReassigned = controllerContext.partitionsBeingReassigned -- partitionsToBeRemoved
+
+      info(s"Removing partitions $partitionsToBeRemoved from the list of reassigned partitions in zookeeper")
+
+      // write the new list to zookeeper
+      if (updatedPartitionsBeingReassigned.isEmpty) {
+        info(s"No more partitions need to be reassigned. Deleting zk path ${ReassignPartitionsZNode.path}")
+        zkClient.deletePartitionReassignment(controllerContext.epochZkVersion)
+      } else {
+        val reassignment = updatedPartitionsBeingReassigned.mapValues(_.newReplicas)
+        try zkClient.setOrCreatePartitionReassignment(reassignment, controllerContext.epochZkVersion)
+        catch {
+          case e: KeeperException => throw new AdminOperationException(e)
+        }
+      }
+
+      controllerContext.partitionsBeingReassigned --= partitionsToBeRemoved
+    }
   }
 
   private def removePartitionsFromPreferredReplicaElection(partitionsToBeRemoved: Set[TopicPartition],
@@ -1555,16 +1555,17 @@ class KafkaController(val config: KafkaConfig, zkClient: KafkaZkClient, time: Ti
       // the `path exists` check for free
       if (zkClient.registerZNodeChangeHandlerAndCheckExistence(partitionReassignmentHandler)) {
         val partitionReassignment = zkClient.getPartitionReassignment
-
+        val partitionsToBeReassigned = partitionReassignment.filterNot(p => controllerContext.partitionsBeingReassigned.contains(p._1))
         // Populate `partitionsBeingReassigned` with all partitions being reassigned before invoking
         // `maybeTriggerPartitionReassignment` (see method documentation for the reason)
-        partitionReassignment.foreach { case (tp, newReplicas) =>
+        partitionsToBeReassigned.foreach { case (tp, newReplicas) =>
           val reassignIsrChangeHandler = new PartitionReassignmentIsrChangeHandler(KafkaController.this, eventManager,
             tp)
           controllerContext.partitionsBeingReassigned.put(tp, ReassignedPartitionsContext(newReplicas, reassignIsrChangeHandler))
         }
 
-        maybeTriggerPartitionReassignment(partitionReassignment.keySet)
+        info("Partitions needed to be reassigned: " + partitionsToBeReassigned)
+        maybeTriggerPartitionReassignment(partitionsToBeReassigned.keySet)
       }
     }
   }
@@ -1749,10 +1750,11 @@ class TopicDeletionFlagHandler(controller: KafkaController, eventManager: Contro
 class PartitionReassignmentHandler(controller: KafkaController, eventManager: ControllerEventManager) extends ZNodeChangeHandler {
   override val path: String = ReassignPartitionsZNode.path
 
-  // Note that the event is also enqueued when the znode is deleted, but we do it explicitly instead of relying on
-  // handleDeletion(). This approach is more robust as it doesn't depend on the watcher being re-registered after
-  // it's consumed during data changes (we ensure re-registration when the znode is deleted).
   override def handleCreation(): Unit = eventManager.put(controller.PartitionReassignment)
+
+  override def handleDeletion(): Unit = eventManager.put(controller.PartitionReassignment)
+
+  override def handleDataChange(): Unit = eventManager.put(controller.PartitionReassignment)
 }
 
 class PartitionReassignmentIsrChangeHandler(controller: KafkaController, eventManager: ControllerEventManager, partition: TopicPartition) extends ZNodeChangeHandler {
